requirements-vis.txt
google-cloud-vision
pyscreenshot
